- Aggiunto un ComponentListener per rendere `PannelloGioco` responsivo. Le carte e le immagini scalano ora proporzionalmente alla dimensione della finestra. - Ridotte le dimensioni massime delle immagini per evitare sbordamenti nel JFrame. - Inibito il bottone "Conferma" della vista quando il `numeroTurno` è diverso da 0 (impedisce input concorrenti). - Spostata l'esecuzione dell'AI avversaria all'interno del Model con l'introduzione di `eseguiTurniBot()`, garantendo un flusso Event-Driven circolare pulito e sicuro.

+++ jbriscola/src/it/com/jbriscola/model/Carta.java
@@ -1,7 +1,7 @@
 package it.com.jbriscola.model;
 
 /**
- * Classe Carta che rappresenta la singola carta da gioco
+ * Classe Carta che rappresenta la singola carta da gioco.
  */
 public class Carta {
 
@@ -12,36 +12,86 @@ public class Carta {
 
     private static final String PATH_CARTE = "assets/carte/";
 
+    /**
+     * Costruttore della classe Carta.
+     * Inizializza la carta con il seme e il valore specificati.
+     * Complessità computazionale: O(1).
+     *
+     * @param seme   il seme della carta
+     * @param valore il valore della carta
+     */
     public Carta(Seme seme, Valore valore) {
         this.seme = seme;
         this.valore = valore;
         pathCarta = getPathCarta();
     }
 
+    /**
+     * Restituisce il seme della carta.
+     * Complessità computazionale: O(1).
+     *
+     * @return il seme della carta
+     */
     public Seme getSeme() {
         return seme;
     }
 
+    /**
+     * Imposta un nuovo seme per la carta.
+     * Complessità computazionale: O(1).
+     *
+     * @param seme il nuovo seme da assegnare
+     */
     public void setSeme(Seme seme) {
         this.seme = seme;
     }
 
+    /**
+     * Restituisce il valore facciale della carta.
+     * Complessità computazionale: O(1).
+     *
+     * @return il valore della carta
+     */
     public Valore getValore() {
         return valore;
     }
 
+    /**
+     * Imposta un nuovo valore per la carta.
+     * Complessità computazionale: O(1).
+     *
+     * @param valore il nuovo valore da assegnare
+     */
     public void setValore(Valore valore) {
         this.valore = valore;
     }
 
-    public int getPuntiCarta() {
+    /**
+     * Restituisce i punti attribuiti alla carta in base al suo valore in Briscola.
+     * Complessità computazionale: O(1).
+     *
+     * @return i punti della carta
+     */
+    public double getPuntiCarta() {
         return valore.getPunti();
     }
 
+    /**
+     * Restituisce il percorso dell'immagine associata alla carta.
+     * Complessità computazionale: O(1), poiché l'operazione di concatenazione è su stringhe costanti/corte.
+     *
+     * @return il percorso del file immagine
+     */
     public String getPathCarta() {
         return PATH_CARTE + valore.name() + "_" + seme.name() + ".png";
     }
 
+    /**
+     * Restituisce una rappresentazione in formato testuale della Carta.
+     * Complessità computazionale: O(1).
+     *
+     * @return una stringa che rappresenta lo stato della carta
+     */
     @Override
     public String toString() {
         return "Carta{" +
